requirements.txt
python-dateutil
fpdf2
openpyxl
tk
matplotlib
matplotlib-inline
Pillow
uuid